fix: remove old button from docs [B1M2-5375]
