--- next-docs/public/examples/button/Secondary.tsx
@@ -1,5 +1,0 @@
-import { Button } from '@heathmont/moon-components';
-
-const Example = () => <Button variant="secondary">Secondary</Button>;
-
-export default Example;
